src/requirements.txt
python_dotenv==0.14.0
PyYAML==5.3.1
jsonschema==3.2.0
pandas==1.1.2
psycopg2==2.8.6
xlrd==1.2.0
chardet==3.0.4